Format code so it passes the linter
